Refactor backend to support separate authentication and market databases
- Updated StreamHandler to use separate AuthDB and MarketDB connections.
- Modified InstrumentStream and FavoritesStream to query from the new MarketDB.
- Changed data types for instrument IDs in models to int64 for consistency.
- Revised database initialization scripts to create auth and ingest schemas.
- Added provisioning script for market database with necessary tables and indexes.
- Updated deployment configurations to include new MARKET_DATABASE_URL secret.

+++ backend/internal/models/models.go
@@ -37,7 +37,7 @@ func (u *User) Public() UserPublic {
 // ---- Instruments ----
 
 type Instrument struct {
-	ID         int     `json:"id"`
+	ID         int64   `json:"id"`
 	Symbol     string  `json:"symbol"`
 	Name       *string `json:"name"`
 	Exchange   *string `json:"exchange"`
@@ -136,7 +136,7 @@ type RefreshRequest struct {
 // ---- Favorites ----
 
 type FavoritesUpdateRequest struct {
-	InstrumentIDs []int `json:"instrument_ids"`
+	InstrumentIDs []int64 `json:"instrument_ids"`
 }
 
 // ---- Market Status ----
